+++ python/GafferUI/UIEditor.py
@@ -1795,4 +1795,5 @@ UIEditor.registerWidgetSetting(
 	_ButtonCodeMetadataWidget,
 )
 UIEditor.registerWidgetMetadata( "Inline", "GafferUI.ButtonPlugValueWidget", "layout:accessory", False )
+UIEditor.registerWidgetMetadata( "Allow Custom Values", "GafferUI.PresetsPlugValueWidget", "presetsPlugValueWidget:allowCustom", False )
 UIEditor.registerWidgetMetadata( "Divider", "*", "divider", False )